+++ example/airports/models.py
@@ -1,29 +1,6 @@
 from django.db import models
 
-from .utils import get_model_fields
-
-
-class PowerTableManager(models.Manager):
-    def get_queryset(self):
-        return (
-            super()
-            .get_queryset()
-            .order_by(*get_model_fields(self.model, option="name"))
-        )
-
-
-class PowerTableModel(models.Model):
-    name = models.CharField(max_length=255, unique=True)
-
-    objects = PowerTableManager()
-
-    class Meta:
-        abstract = True
-
-    def __str__(self):
-        return self.name
-
-    __repr__ = __str__
+from example.powertables.models import PowerTableModel
 
 
 class Iata(PowerTableModel):