+++ example/project/settings.py
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "powertables.apps.PowertablesConfig",
     "airports.apps.AirportsConfig",
 ]