+++ CHANGELOG.md
@@ -12,3 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add affiliation section to readme
 - Add licensing
 - Add spike (as example)
+
+### Changed
+
+- Extract generic powertables app in example